Update template: fix template build, add readme, clean up MainActivity code.
